refactor(kernel): move proxy resolution to an untagged file so its test runs in CI (M4)
proxyForEndpoint / proxyForEndpointFunc are pure Go (only config + net/http/url,
no kernel C symbol), but lived in the cgo-tagged kernel_backend.go — so the
now-meaningful TestProxyForEndpoint (four asserted branches) was inert under the
only CI job (CGO_ENABLED=0). Split them into an untagged kernel_proxy.go +
kernel_proxy_test.go, mirroring the arrowscan/decimalfmt move; newKernelBackend
stays gated and calls proxyForEndpoint. The proxy test now runs in the default
matrix without a kernel lib.
